requirements.txt
streamlit==1.31.0
pandas==2.1.3
PyPDF2==3.0.1
python-docx==1.0.1
pillow==10.2.0
scikit-learn==1.3.2
numpy==1.26.2
matplotlib==3.8.2
openai==1.12.0
requests==2.31.0
opencv-python==4.8.0.76
pdfplumber==0.10.3